+++ optimize_workspace.sh
@@ -6,6 +6,11 @@ command -v shellcheck >/dev/null && shellcheck "$0"
 
 export PATH="$PATH:/root/.cargo/bin"
 
+# Suffix for non-Intel built artifacts
+MACHINE=$(uname -m)
+SUFFIX=${MACHINE#x86_64}
+SUFFIX=${SUFFIX:+-$SUFFIX}
+
 rustup toolchain list
 cargo --version
 
@@ -27,10 +32,9 @@ TMPARTIFACTS=$(mktemp -p "$(pwd)" -d artifacts.XXXXXX)
   cd "$TMPARTIFACTS"
 
   for WASM in ../target/wasm32-unknown-unknown/release/*.wasm; do
-    echo "Optimizing $WASM ..."
-    BASE=$(basename "$WASM")
-    wasm-opt -Os -o "$BASE" "$WASM"
-    chmod -x "$BASE"
+    NAME=$(basename "$WASM" .wasm)${SUFFIX}.wasm
+    echo "Optimizing $NAME ..."
+    wasm-opt -Os "$WASM" -o "$NAME"
   done
   echo "Moving wasm files ..."
   mv ./*.wasm ../artifacts